First cut at inline filtering
This adds an implementation of in-line filtering (alongside the
current broken post-process filters).

This introduces a couple of new classes, the important ones are

  RegexFilter
  FilterIO

The first is a simple regular-expression based search and replace
with the same behaviour as the existing doFileSub()/doCmdSub()
regular expression filters. The pattern and replacement string
are set when the object is initialised and are used to compile
an re object. The filter's filter() method takes a buffer of text
and returns the filtered text

To make this work with Tarfile.addfile() this uses FilterIO as a
wrapper to apply the filter to data as it is read.

This would allow us to filter files as they are read from disk and
without buffering the entire file in memory but has a couple of
drawbacks and complications:

 - filters cannot change the size of files
   The tar header is written to the tar file before the data; this
   means that the size must be known before reading commences.

   With the current interface this has some real problems in the
   case that the substituted string is longer than the original.
   To maintain consistency something has to be truncated and this
   is hard to do correctly with the unstructured search/replace.

This isn't currently fit for master and may need to be replaced
with a /tmp-based temporary tree.